refactor(query): refactor test style
