+++ src/query/expression/tests/it/block.rs
@@ -1,10 +1,15 @@
+use databend_common_column::buffer::Buffer;
 use databend_common_expression::block_debug::box_render;
 use databend_common_expression::types::number::NumberScalar;
 use databend_common_expression::types::string::StringColumnBuilder;
 use databend_common_expression::types::AnyType;
+use databend_common_expression::types::ArrayColumn;
 use databend_common_expression::types::DataType;
 use databend_common_expression::types::Int32Type;
+use databend_common_expression::types::Int64Type;
 use databend_common_expression::types::NumberDataType;
+use databend_common_expression::types::StringType;
+use databend_common_expression::types::ValueType;
 use databend_common_expression::BlockEntry;
 use databend_common_expression::Column;
 use databend_common_expression::DataField;
@@ -75,4 +80,26 @@ fn test_block_entry_memory_size() {
     let scalar_str = Scalar::String("abc".to_string());
     let entry = BlockEntry::new(DataType::String, Value::<AnyType>::Scalar(scalar_str));
     assert_eq!(3, entry.memory_size());
+
+    let col = StringType::from_data((0..10).map(|x| x.to_string()).collect::<Vec<_>>());
+    assert_eq!(col.memory_size(), 10 + 10 * 12);
+
+    let array = ArrayColumn::<Int64Type>::new(
+        Buffer::from_iter(0..10i64),
+        Buffer::from(vec![0u64, 1, 3, 6, 10]),
+    );
+    let total_memory_size = array.memory_size(); // 10 * 8 + 5 * 8 = 80 + 40 = 120
+    let expected = array
+        .iter()
+        .map(|x| Int64Type::column_memory_size(&x))
+        .sum::<usize>()
+        + (array.len() + 1) * 8;
+    assert_eq!(total_memory_size, expected);
+
+    let array2 = array.slice(0..2);
+    let array3 = array.slice(2..4);
+    assert_eq!(
+        total_memory_size,
+        array2.memory_size() + array3.memory_size() - 8
+    );
 }
