+++ src/query/formats/src/field_encoder/helpers/escape.rs
@@ -91,3 +91,106 @@ pub fn write_tsv_escaped_string(bytes: &[u8], buf: &mut Vec<u8>, field_delimiter
         buf.extend_from_slice(&bytes[start..]);
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::field_encoder::helpers::write_json_string;
+    use crate::field_encoder::helpers::write_tsv_escaped_string;
+    use crate::field_encoder::write_csv_string;
+
+    #[test]
+    fn test_escape() {
+        {
+            let mut buf = vec![];
+            write_tsv_escaped_string(b"\0\n\r\t\\\'", &mut buf, b'\t');
+            assert_eq!(&buf, b"\\0\\n\\r\\t\\\\\'")
+        }
+
+        {
+            let mut buf = vec![];
+            write_tsv_escaped_string(b"\n123\n456\n", &mut buf, b'\t');
+            assert_eq!(&buf, b"\\n123\\n456\\n")
+        }
+
+        {
+            let mut buf = vec![];
+            write_tsv_escaped_string(b"123\n", &mut buf, b'\t');
+            assert_eq!(&buf, b"123\\n")
+        }
+
+        {
+            let mut buf = vec![];
+            write_tsv_escaped_string(b"\n123", &mut buf, b'\t');
+            assert_eq!(&buf, b"\\n123")
+        }
+
+        {
+            let mut buf = vec![];
+            write_tsv_escaped_string(b"\n,23", &mut buf, b',');
+            assert_eq!(&buf, b"\\n\\,23")
+        }
+    }
+
+    #[test]
+    fn test_json_escape() {
+        let basic = b"\0\n\r\t\'/\"\\";
+        {
+            let mut buf = vec![];
+            write_json_string(basic, &mut buf, false, true);
+            assert_eq!(&buf, b"\\u0000\\n\\r\\t\'\\/\\\"\\\\")
+        }
+
+        {
+            let mut buf = vec![];
+            write_json_string(basic, &mut buf, false, false);
+            assert_eq!(&buf, b"\\u0000\\n\\r\\t\'/\\\"\\\\")
+        }
+
+        {
+            let mut buf = vec![];
+            write_json_string(basic, &mut buf, true, true);
+            assert_eq!(&buf, b"\\u0000\\n\\r\\t\'\\/\"\"\\\\")
+        }
+
+        {
+            let mut buf = vec![];
+            write_json_string(b"\n123\n456\n", &mut buf, true, true);
+            assert_eq!(&buf, b"\\n123\\n456\\n")
+        }
+
+        {
+            let mut buf = vec![];
+            write_json_string(b"123\n", &mut buf, true, true);
+            assert_eq!(&buf, b"123\\n")
+        }
+
+        {
+            let mut buf = vec![];
+            write_json_string(b"\n123", &mut buf, true, true);
+            assert_eq!(&buf, b"\\n123")
+        }
+
+        {
+            let mut buf = vec![];
+            write_json_string(b"\n123", &mut buf, true, true);
+            assert_eq!(&buf, b"\\n123")
+        }
+
+        {
+            let s = "123\u{2028}\u{2029}abc";
+            let mut buf = vec![];
+            write_json_string(s.as_bytes(), &mut buf, true, true);
+            assert_eq!(&buf, b"123\\u2028\\u2029abc")
+        }
+    }
+
+    #[test]
+    fn test_csv_string() {
+        {
+            let s = "a\"\nb";
+            let mut buf = vec![];
+            write_csv_string(s.as_bytes(), &mut buf, b'"');
+            assert_eq!(&buf, b"\"a\"\"\nb\"")
+        }
+    }
+}
