--- src/query/formats/tests/it/testdata/fast_values.txt
@@ -1,82 +1,0 @@
----------- Input Data ----------
-"(0, 1, 2), (3,4,5)"
----------- Input Column Types ----------
-[Number(Int16), Number(Int16), Number(Int16)]
----------- Output ---------
-+----------+----------+----------+
-| Column 0 | Column 1 | Column 2 |
-+----------+----------+----------+
-| 0        | 1        | 2        |
-| 3        | 4        | 5        |
-+----------+----------+----------+
-
-
----------- Input Data ----------
-"(0, 1, 2), (3,4,5), "
----------- Input Column Types ----------
-[Number(Int16), Number(Int16), Number(Int16)]
----------- Output ---------
-BadDataValueType. Code: 1010, Text = Must start with parentheses.
-
-
----------- Input Data ----------
-"('', '', '')"
----------- Input Column Types ----------
-[Number(Int16), Number(Int16), Number(Int16)]
----------- Output ---------
-Unimplemented. Code: 1002, Text = fallback.
-
-
----------- Input Data ----------
-"( 1, '', '2022-10-01')"
----------- Input Column Types ----------
-[Number(Int16), String, Date]
----------- Output ---------
-+----------+----------+--------------+
-| Column 0 | Column 1 | Column 2     |
-+----------+----------+--------------+
-| 1        | ''       | '2022-10-01' |
-+----------+----------+--------------+
-
-
----------- Input Data ----------
-"(1, 2, 3), (1, 1, 1), (1, 1, 1);"
----------- Input Column Types ----------
-[Number(Int16), Number(Int16), Number(Int16)]
----------- Output ---------
-+----------+----------+----------+
-| Column 0 | Column 1 | Column 2 |
-+----------+----------+----------+
-| 1        | 2        | 3        |
-| 1        | 1        | 1        |
-| 1        | 1        | 1        |
-+----------+----------+----------+
-
-
----------- Input Data ----------
-"(1, 2, 3), (1, 1, 1), (1, 1, 1);  "
----------- Input Column Types ----------
-[Number(Int16), Number(Int16), Number(Int16)]
----------- Output ---------
-+----------+----------+----------+
-| Column 0 | Column 1 | Column 2 |
-+----------+----------+----------+
-| 1        | 2        | 3        |
-| 1        | 1        | 1        |
-| 1        | 1        | 1        |
-+----------+----------+----------+
-
-
----------- Input Data ----------
-"(1.2, -2.9, 3.55), (3.12e2, 3.45e+3, -1.9e-3);"
----------- Input Column Types ----------
-[Number(Int16), Number(Int16), Number(Int16)]
----------- Output ---------
-+----------+----------+----------+
-| Column 0 | Column 1 | Column 2 |
-+----------+----------+----------+
-| 1        | -3       | 4        |
-| 312      | 3450     | 0        |
-+----------+----------+----------+
-
-
